+++ build_runner/CHANGELOG.md
@@ -15,6 +15,7 @@
   crashing.
 - Bug fix: fix `dart run build_runner test` to correctly pass arguments after
   `--` to the test process.
+- Bug fix: fix crash if a resolved Dart source contains invalid utf8.
 - Require Dart 3.8.0.
 
 ## 2.15.0